Revert DependencyContext changes untill dotnet-cli packages get sorted out
