+++ src/Microsoft.AspNet.Mvc.Core/project.json
@@ -21,7 +21,6 @@
       "version": "1.0.0-*"
     },
     "Microsoft.Extensions.PlatformAbstractions": "1.0.0-*",
-    "Microsoft.Extensions.DependencyModel": "1.0.0-*",
     "Microsoft.Extensions.ClosedGenericMatcher.Sources": {
       "version": "1.0.0-*",
       "type": "build"
@@ -45,7 +44,6 @@
   "frameworks": {
     "net451": {
       "frameworkAssemblies": {
-        "System.Reflection": "",
         "System.Runtime": ""
       }
     },
